Zowe copybook resolving refactor
