+++ clients/cobol-lsp-vscode-extension/src/services/copybookLibs/DatasetLib.ts
@@ -2,7 +2,6 @@ import { DATASET } from "../../constants";
 import * as vscode from "vscode";
 import CopybookLib from "./CopybookLib";
 import { CopybookLibs } from "../ProcessorGroupsLoader";
-import { ProfileUtils } from "../util/ProfileUtils";
 import { externalApis } from "../ExternalAPIsService";
 import { ZoweLib } from "./ZoweLib";
 import { MainframeRemoteLocation } from "../copybook/downloader/DownloadUtil";
@@ -19,7 +18,6 @@ export class DatasetLib extends ZoweLib implements CopybookLib {
     const libs = [];
     for (const config of configs) {
       if (typeof config === "object" && DATASET in config) {
-        // if (hasMember(config, 'DATASET') && typeof config[DATASET] === 'string') {
         libs.push(new DatasetLib(config.dataset, config.profile));
       }
     }
@@ -34,41 +32,37 @@ export class DatasetLib extends ZoweLib implements CopybookLib {
     copybookName: string,
     documentUri: vscode.Uri,
   ): Promise<vscode.Uri | undefined> {
-    if (!this.profile) {
-      this.profile = ProfileUtils.getProfileNameForCopybook(
-        documentUri,
-        externalApis.explorerApi,
-      );
-    }
+    const profile = this.getProfile(documentUri);
 
     if (!(await this.configCheck(documentUri))) {
       return;
     }
 
-    return await externalApis.dsnService?.resolveCopybookUri(
-      this.profile ?? "profile",
+    const member = await externalApis.dsnService?.hasMember(
+      profile,
       this.dsn,
       copybookName,
     );
+
+    if (member) {
+      return vscode.Uri.parse(
+        `zowe-dsn:/${profile}/${this.dsn}/${member.name}${member.extension ? member.extension : ""}`,
+      );
+    }
   }
 
   async listCopybooks(
     documentUri: vscode.Uri,
     _outputChannel?: vscode.OutputChannel,
   ): Promise<string[]> {
-    if (!this.profile) {
-      this.profile = ProfileUtils.getProfileNameForCopybook(
-        documentUri,
-        externalApis.explorerApi,
-      );
-    }
+    const profile = this.getProfile(documentUri);
 
     if (!(await this.configCheck(documentUri))) {
       return [];
     }
 
     const members = await externalApis.dsnService?.getAllMembers(
-      this.profile ?? "profile",
+      profile,
       this.dsn,
     );
